Add example TileMap

+++ include/LDL/Graphics/Texture.hpp
@@ -13,7 +13,7 @@ namespace LDL
 		class Texture : public LDL::Core::FastPimpl
 		{
 		public:
-			Texture(const Point2u& size, size_t bytesPerPixel, uint8_t* pixels);
+			Texture(const Point2u& size, uint8_t bytesPerPixel, uint8_t* pixels);
 			~Texture();
 			const Point2u& Size();
 			size_t Id();
